% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAPE.R
\name{auto_anno}
\alias{auto_anno}
\title{auto_anno}
\usage{
auto_anno(
  input_dir = NULL,
  output_dir = NULL,
  output_plot = TRUE,
  output_plot_format = "png",
  n_features = 2000,
  qc_all = 25,
  qc_immune = 5,
  doublet_reserve = FALSE
)
}
\arguments{
\item{input_dir}{todo}

\item{output_dir}{todo}

\item{output_plot}{todo}

\item{output_plot_format}{todo}

\item{n_features}{todo}

\item{qc_all}{todo}

\item{qc_immune}{todo}

\item{doublet_reserve}{todo}
}
\description{
todo
}
\details{
auto_anno
}
