% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scUmaper.R
\name{run_custom_scumaper}
\alias{run_custom_scumaper}
\title{run_custom_scumaper}
\usage{
run_custom_scumaper(
  input_seurat = NULL,
  input_genelist = NULL,
  output_dir = NULL,
  output_plot = TRUE,
  output_plot_format = "png",
  n_features = 3000,
  doublet_reserve = FALSE
)
}
\arguments{
\item{input_seurat}{Seurat object to remove doublets and annotate.}

\item{input_genelist}{A list of several vectors, which named by target clusters and composed of corresponding marker genes.}

\item{output_dir}{Path to store output files, defaults to create a new folder named "scUmaper_output" under working directory.}

\item{output_plot}{Whether outputting UMAP plots or not, defaults to TRUE.}

\item{output_plot_format}{Format of outputting UMAP plots, should be "png" or "pdf", defaults to "png".}

\item{n_features}{Number of feature genes when clustering, defaults to 3000.}

\item{doublet_reserve}{Whether reserving doublet cells as another Seurat object, defaults to FALSE.}
}
\description{
run_custom_scumaper() is a function for executing scUmaper from processed Seurat objects, while users can customize the marker genes to use.
}
\details{
run_custom_scumaper
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
run_scumaper(input_seurat = scRNA, input_genelist = list("T_cell" = c("CD3D", "CD3E"), "Plasma_cell" = c("MZB1", "JCHAIN")), output_dir = 'path/to/output')
\dontshow{\}) # examplesIf}
}
