% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scUmaper.R
\name{run_scumaper}
\alias{run_scumaper}
\title{run_scumaper}
\usage{
run_scumaper(
  input_dir = NULL,
  output_dir = NULL,
  output_plot = TRUE,
  output_plot_format = "png",
  n_features = 3000,
  qc_all = 25,
  qc_immune = 5,
  doublet_reserve = FALSE
)
}
\arguments{
\item{input_dir}{Path of the folder storing raw single cell gene expression matrices. This path should include several folders named by the samples, which should include 3 files named "barcodes.tsv.gz", "features.tsv.gz", and "matrix.mtx.gz".}

\item{output_dir}{Path to store output files, defaults to create a new folder named "scUmaper_output" under working directory.}

\item{output_plot}{Whether outputting UMAP plots or not, defaults to TRUE.}

\item{output_plot_format}{Format of outputting UMAP plots, should be "png" or "pdf", defaults to "png".}

\item{n_features}{Number of feature genes when clustering, defaults to 3000.}

\item{qc_all}{Quality control standard of all cells, the cells above how many percentage of mitochondrial gene expression will be removed, defaults to 25.}

\item{qc_immune}{Quality control standard of immune cells, the cells above how many percentage of mitochondrial gene expression will be removed, defaults to 5.}

\item{doublet_reserve}{Whether reserving doublet cells as another Seurat object, defaults to FALSE.}
}
\description{
run_scumaper() is a function for executing scUmaper from raw single cell gene expression matrices.
}
\details{
run_scumaper
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
run_scumaper(input_dir = 'path/to/input', output_dir = 'path/to/output')
\dontshow{\}) # examplesIf}
}
